+++ crypto/pbkdf2.h
@@ -24,6 +24,7 @@
 #ifndef __PBKDF2_H__
 #define __PBKDF2_H__
 
+#include <stddef.h>
 #include <stdint.h>
 #include "sha2.h"
 
@@ -44,23 +45,23 @@ typedef struct _PBKDF2_HMAC_SHA512_CTX {
 } PBKDF2_HMAC_SHA512_CTX;
 
 void pbkdf2_hmac_sha256_Init(PBKDF2_HMAC_SHA256_CTX *pctx, const uint8_t *pass,
-                             int passlen, const uint8_t *salt, int saltlen,
-                             uint32_t blocknr);
+                             size_t passlen, const uint8_t *salt,
+                             size_t saltlen, uint32_t blocknr);
 void pbkdf2_hmac_sha256_Update(PBKDF2_HMAC_SHA256_CTX *pctx,
                                uint32_t iterations);
 void pbkdf2_hmac_sha256_Final(PBKDF2_HMAC_SHA256_CTX *pctx, uint8_t *key);
-void pbkdf2_hmac_sha256(const uint8_t *pass, int passlen, const uint8_t *salt,
-                        int saltlen, uint32_t iterations, uint8_t *key,
-                        int keylen);
+void pbkdf2_hmac_sha256(const uint8_t *pass, size_t passlen,
+                        const uint8_t *salt, size_t saltlen,
+                        uint32_t iterations, uint8_t *key, size_t keylen);
 
 void pbkdf2_hmac_sha512_Init(PBKDF2_HMAC_SHA512_CTX *pctx, const uint8_t *pass,
-                             int passlen, const uint8_t *salt, int saltlen,
-                             uint32_t blocknr);
+                             size_t passlen, const uint8_t *salt,
+                             size_t saltlen, uint32_t blocknr);
 void pbkdf2_hmac_sha512_Update(PBKDF2_HMAC_SHA512_CTX *pctx,
                                uint32_t iterations);
 void pbkdf2_hmac_sha512_Final(PBKDF2_HMAC_SHA512_CTX *pctx, uint8_t *key);
-void pbkdf2_hmac_sha512(const uint8_t *pass, int passlen, const uint8_t *salt,
-                        int saltlen, uint32_t iterations, uint8_t *key,
-                        int keylen);
+void pbkdf2_hmac_sha512(const uint8_t *pass, size_t passlen,
+                        const uint8_t *salt, size_t saltlen,
+                        uint32_t iterations, uint8_t *key, size_t keylen);
 
 #endif
